fix(site): ThemeToggle moves to site/, ProjectRow guard simplified, mobile sheet inert trap
- Move ThemeToggle from app/components/blog/ to app/components/site/.
  The toggle is site-wide chrome (used by TopBar, mounted on every
  page) so the file lived in the wrong namespace — site/ depending on
  blog/ was a backwards dependency.
- Simplify ProjectRow's redundant guard. 'hasScreenshot' was set from
  Boolean(screenshotUrl && targetUrl), then the JSX repeated
  'hasScreenshot && screenshotUrl && targetUrl ? ...'. Replaced with
  an explicit 'showScreenshot = screenshotUrl != null && targetUrl !=
  null' that narrows the union types so the JSX guard is just
  'showScreenshot ? ...'.
- Add inert on <main> when the mobile sheet is open. Tab now can't
  escape the sheet into the page underneath. The topbar (containing
  the trigger and sheet) lives outside <main> so it stays interactive.
  Effect cleans up on close and on unmount.
